Fix method name matching file name #1510
