+++ Data_Engine/Data_Engine.csproj
@@ -95,6 +95,8 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Query\Axes.cs" />
     <Compile Include="Query\AxisExists.cs" />
+    <Compile Include="Query\FirstValueAbove.cs" />
+    <Compile Include="Query\FirstValueAt.cs" />
     <Compile Include="Query\Linq\FilterBy.cs" />
     <Compile Include="Query\Linq\OrderBy.cs" />
     <Compile Include="Query\CloseToPoint.cs" />
@@ -108,8 +110,8 @@
     <Compile Include="Query\Peek.cs" />
     <Compile Include="Query\SquareDistance.cs" />
     <Compile Include="Query\SubMatrixData.cs" />
-    <Compile Include="Query\ValueAbove.cs" />
-    <Compile Include="Query\ValueAt.cs" />
+    <Compile Include="Query\ValuesAbove.cs" />
+    <Compile Include="Query\ValuesAt.cs" />
     <Compile Include="Query\Values.cs" />
   </ItemGroup>
   <ItemGroup>
